requirements.txt
# ───────────────────────────────────────────
# AccessMesh-AI — Python Dependencies
# ───────────────────────────────────────────

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Rate limiting
slowapi>=0.1.9

# JWT authentication
python-jose[cryptography]>=3.3.0

# Password hashing (bcrypt)
passlib[bcrypt]>=1.7.4
bcrypt>=3.1.0,<4.0

# Azure Web PubSub
azure-messaging-webpubsubservice>=1.1.0
azure-core>=1.30.0

# Azure Service Bus
azure-servicebus>=7.12.0

# Azure Speech Service
azure-cognitiveservices-speech>=1.37.0

# Azure Cosmos DB (session + message persistence)
# aiohttp is required by azure-cosmos async client (azure.cosmos.aio)
azure-cosmos>=4.7.0
aiohttp>=3.9.0

# Azure Content Safety
azure-ai-contentsafety>=1.0.0

# Azure AI Translator
azure-ai-translation-text>=1.0.0

# Azure Application Insights (OpenTelemetry)
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0

# WebSocket support for FastAPI
websockets>=12.0

# Validation / Serialization
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0

# HTTP client (used internally by Azure SDK and agents)
httpx>=0.27.0

# Environment variable loading
python-dotenv>=1.0.0

# Azure identity (for managed identity / AAD auth)
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Logging / utilities
loguru>=0.7.2

# Multipart form data (file uploads — required by /speech/transcribe)
python-multipart>=0.0.9
